+++ src/common/context.h
@@ -36,7 +36,7 @@ struct context_metadata {
   UT_hash_handle hh;  // makes this structure hashable
 };
 
-#define NUM_CONTEXT_MODEL_FLAGS 11
+#define NUM_CONTEXT_MODEL_FLAGS 12
 // See docs/developer-guide/cli-options.md for details on how to add a new
 // Context entry
 struct Context {
@@ -53,7 +53,7 @@ struct Context {
   int nitrogenCycle;
   int anaerobic;
   int flooding;
-
+  int carbonSaturation;
   // IF ADDING A NEW MODEL FLAG, update NUM_CONTEXT_MODEL_FLAGS above and
   // relevant code in restart.c