+++ docs/CHANGELOG.md
@@ -30,6 +30,7 @@ sections to include in release notes:
 - `leafon` and `leafoff` events for tracking phenological transitions (#326)
 - `leafon` limited by available carbon and nitrogen; N storage pool; N resorption on `leafoff` (#337)
 - New required parameter `leafOnReallocFrac` to control how much of wood and coarse root carbon is reallocated to leaves on `leafon` (#337)
+- `carbonSaturation` cli option and calculation of soil and litter carbon pools that observes soil carbon saturating behavior (#301)
 
 ### Fixed
 
@@ -38,6 +39,8 @@ sections to include in release notes:
 - Renamed the CLI option `--file-name` to `--file-prefix` for clarity while keeping `--file-name` as a backward-compatible alias (#320)
 
 ### Removed
+
+- Schema version from restart checkpoint format; model version is sufficient (#338)
 
 ### Git SHA